seperate sockets from server
